+++ package.json
@@ -12,6 +12,7 @@
     "react": "^17.0.2",
     "react-color": "^2.19.3",
     "react-dom": "^17.0.2",
+    "react-router-dom": "^6.0.2",
     "react-scripts": "4.0.3",
     "reactcss": "^1.2.3",
     "web-vitals": "^1.1.2"
