+++ src/Routes/Canvas/Canvas.css
@@ -1,4 +1,4 @@
-/* .App {
+.canvas {
     display: grid;
     grid-template-rows: 0.1fr 1fr;
 }
@@ -13,4 +13,4 @@
 
 #paint-text {
     margin-left: 17em;
-} */
+}
